requirements.txt
openai>=1.0.0
httpx
tqdm
numpy
pyyaml